# cuda4dnn(conv): use cudnn for conv + bias + relu fusion