--- a/modules/dnn/src/cuda4dnn/primitives/convolution.hpp
+++ b/modules/dnn/src/cuda4dnn/primitives/convolution.hpp
@@ -204,6 +204,21 @@ namespace cv { namespace dnn { namespace cuda4dnn {
             params.dilation = dilations;
             params.groups = config.groups;
 
+            /* check if we can perform fused convolution using cudnn */
+            params.activation_type = csl::Convolution<T>::ActivationType::IDENTITY;
+            fusion_location = InternalFusionLocation::NATIVE;
+            if (!biasTensor.empty() &&
+                biasTensor.size() == output_feature_maps && /* cuDNN requirement */
+                config.activation_type == ConvolutionConfiguration::ActivationType::RELU &&
+                config.relu_negative_slope == 0.0)
+            {
+                fusion_location = InternalFusionLocation::CUDNN;
+                auto bias_shape = std::vector<std::size_t>(rank, 1);
+                bias_shape[1] = output_feature_maps;
+                params.bias_shape = bias_shape;
+                params.activation_type = csl::Convolution<T>::ActivationType::RELU;
+            }
+
             convoluter = csl::Convolution<T>(cudnnHandle, params);
 
             activation = config.activation_type;
@@ -216,7 +231,8 @@ namespace cv { namespace dnn { namespace cuda4dnn {
                 activation = ConvolutionConfiguration::ActivationType::IDENTITY;
 
             csl::WorkspaceBuilder builder;
-            if (!transformed_shape.empty()) {
+            if (!transformed_shape.empty())
+            {
                 auto& shape = transformed_shape;
                 auto sz = std::accumulate(std::begin(shape), std::end(shape), 1, std::multiplies<std::size_t>());
                 builder.require<T>(sz);
@@ -248,65 +264,72 @@ namespace cv { namespace dnn { namespace cuda4dnn {
             auto output_wrapper = outputs[0].dynamicCast<wrapper_type>();
             auto output = output_wrapper->getSpan();
 
-            convoluter.convolve(output, input, filtersTensor, allocator.get_instance());
-            if (!biasTensor.empty())
+            if (fusion_location == InternalFusionLocation::CUDNN)
             {
-                std::size_t inner_size = output.size_range(2, output.rank());
-                switch(activation)
-                {
-                    case ConvolutionConfiguration::ActivationType::IDENTITY:
-                        kernels::biasN<T>(stream, output, output, inner_size, biasTensor);
-                        break;
-                    case ConvolutionConfiguration::ActivationType::RELU:
-                        kernels::biasN_relu_inplace<T>(stream, output, inner_size, biasTensor, relu_negative_slope);
-                        break;
-                    case ConvolutionConfiguration::ActivationType::CLIPPED_RELU:
-                        kernels::biasN_clipped_relu_inplace<T>(stream, output, inner_size, biasTensor, crelu_floor, crelu_ceil);
-                        break;
-                    case ConvolutionConfiguration::ActivationType::POWER:
-                        kernels::biasN_power_inplace<T>(stream, output, inner_size, biasTensor, power_exp);
-                        break;
-                    case ConvolutionConfiguration::ActivationType::TANH:
-                        kernels::biasN_tanh_inplace<T>(stream, output, inner_size, biasTensor);
-                        break;
-                    case ConvolutionConfiguration::ActivationType::SIGMOID:
-                        kernels::biasN_sigmoid_inplace<T>(stream, output, inner_size, biasTensor);
-                        break;
-                    case ConvolutionConfiguration::ActivationType::SWISH:
-                        kernels::biasN_swish_inplace<T>(stream, output, inner_size, biasTensor);
-                        break;
-                    case ConvolutionConfiguration::ActivationType::MISH:
-                        kernels::biasN_mish_inplace<T>(stream, output, inner_size, biasTensor);
-                        break;
-                }
+                convoluter.convolve_with_bias_activation(output, input, filtersTensor, biasTensor, allocator.get_instance());
             }
             else
             {
-                switch(activation)
+                convoluter.convolve(output, input, filtersTensor, allocator.get_instance());
+                if (!biasTensor.empty())
                 {
-                    case ConvolutionConfiguration::ActivationType::IDENTITY:
-                        break;
-                    case ConvolutionConfiguration::ActivationType::RELU:
-                        kernels::relu<T>(stream, output, output, relu_negative_slope);
-                        break;
-                    case ConvolutionConfiguration::ActivationType::CLIPPED_RELU:
-                        kernels::clipped_relu<T>(stream, output, output, crelu_floor, crelu_ceil);
-                        break;
-                    case ConvolutionConfiguration::ActivationType::POWER:
-                        kernels::power<T>(stream, output, output, power_exp, 1.0, 0.0);
-                        break;
-                    case ConvolutionConfiguration::ActivationType::TANH:
-                        kernels::tanh<T>(stream, output, output);
-                        break;
-                    case ConvolutionConfiguration::ActivationType::SIGMOID:
-                        kernels::sigmoid<T>(stream, output, output);
-                        break;
-                    case ConvolutionConfiguration::ActivationType::SWISH:
-                        kernels::swish<T>(stream, output, output);
-                        break;
-                    case ConvolutionConfiguration::ActivationType::MISH:
-                        kernels::mish<T>(stream, output, output);
-                        break;
+                    std::size_t inner_size = output.size_range(2, output.rank());
+                    switch(activation)
+                    {
+                        case ConvolutionConfiguration::ActivationType::IDENTITY:
+                            kernels::biasN<T>(stream, output, output, inner_size, biasTensor);
+                            break;
+                        case ConvolutionConfiguration::ActivationType::RELU:
+                            kernels::biasN_relu_inplace<T>(stream, output, inner_size, biasTensor, relu_negative_slope);
+                            break;
+                        case ConvolutionConfiguration::ActivationType::CLIPPED_RELU:
+                            kernels::biasN_clipped_relu_inplace<T>(stream, output, inner_size, biasTensor, crelu_floor, crelu_ceil);
+                            break;
+                        case ConvolutionConfiguration::ActivationType::POWER:
+                            kernels::biasN_power_inplace<T>(stream, output, inner_size, biasTensor, power_exp);
+                            break;
+                        case ConvolutionConfiguration::ActivationType::TANH:
+                            kernels::biasN_tanh_inplace<T>(stream, output, inner_size, biasTensor);
+                            break;
+                        case ConvolutionConfiguration::ActivationType::SIGMOID:
+                            kernels::biasN_sigmoid_inplace<T>(stream, output, inner_size, biasTensor);
+                            break;
+                        case ConvolutionConfiguration::ActivationType::SWISH:
+                            kernels::biasN_swish_inplace<T>(stream, output, inner_size, biasTensor);
+                            break;
+                        case ConvolutionConfiguration::ActivationType::MISH:
+                            kernels::biasN_mish_inplace<T>(stream, output, inner_size, biasTensor);
+                            break;
+                    }
+                }
+                else
+                {
+                    switch(activation)
+                    {
+                        case ConvolutionConfiguration::ActivationType::IDENTITY:
+                            break;
+                        case ConvolutionConfiguration::ActivationType::RELU:
+                            kernels::relu<T>(stream, output, output, relu_negative_slope);
+                            break;
+                        case ConvolutionConfiguration::ActivationType::CLIPPED_RELU:
+                            kernels::clipped_relu<T>(stream, output, output, crelu_floor, crelu_ceil);
+                            break;
+                        case ConvolutionConfiguration::ActivationType::POWER:
+                            kernels::power<T>(stream, output, output, power_exp, 1.0, 0.0);
+                            break;
+                        case ConvolutionConfiguration::ActivationType::TANH:
+                            kernels::tanh<T>(stream, output, output);
+                            break;
+                        case ConvolutionConfiguration::ActivationType::SIGMOID:
+                            kernels::sigmoid<T>(stream, output, output);
+                            break;
+                        case ConvolutionConfiguration::ActivationType::SWISH:
+                            kernels::swish<T>(stream, output, output);
+                            break;
+                        case ConvolutionConfiguration::ActivationType::MISH:
+                            kernels::mish<T>(stream, output, output);
+                            break;
+                    }
                 }
             }
         }
@@ -326,6 +349,11 @@ namespace cv { namespace dnn { namespace cuda4dnn {
 
         ConvolutionConfiguration::ActivationType activation;
         float relu_negative_slope, crelu_floor, crelu_ceil, power_exp;
+
+        enum class InternalFusionLocation {
+            CUDNN,
+            NATIVE
+        } fusion_location;
     };
 
 }}} /* namespace cv::dnn::cuda4dnn */