### This pullrequest changes
cuDNN supports conv + bias + relu fusion in some configurations. This PR enables this cuDNN optimized path whenever possible.

**Device:** GTX 1050

Model                               | without this patch | with this patch
--------------------------------- | ----------------------- | -------------------
Inception v2 MaskRCNN (1024x1024) | 187.6ms               | **180.3ms**