@tompollok 2019-12-29
Which card was that? 1050/1060?