@YashasSamaga 2019-12-29
@tompollok GTX 1050 and note that I have increased the input image size. It's the same image which was being used earlier but it has been resized to 1024 x 1024.